+++ src/main/java/backend/capstone/domain/user/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package backend.capstone.domain.user.mapper;
 
+import backend.capstone.auth.util.OAuthAttributes;
 import backend.capstone.domain.user.entity.User;
-import backend.capstone.global.util.OAuthAttributes;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
